Move BlendShapeAvatar.AddBlendShapeClip
