+++ Scripts/BlendShape/BlendShapeAvatar.cs
@@ -3,6 +3,10 @@ using System.Linq;
 using System;
 using System.Collections.Generic;
 using UniGLTF;
+using System.IO;
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace VRM
 {
@@ -36,6 +40,31 @@ namespace VRM
                 Debug.LogFormat("{0}", clip.name);
             }
             Clips = Clips.OrderBy(x => BlendShapeKey.CreateFrom(x)).ToList();
+        }
+
+        public void AddBlendShapeClip()
+        {
+            var dir = Path.GetDirectoryName(AssetDatabase.GetAssetPath(this));
+            var path = EditorUtility.SaveFilePanel(
+                           "Create BlendShapeClip",
+                           dir,
+                           string.Format("BlendShapeClip#{0}.asset", Clips.Count),
+                           "asset");
+            if (string.IsNullOrEmpty(path))
+            {
+                return;
+            }
+            path = path.ToUnityRelativePath();
+            //Debug.LogFormat("{0}", path);
+            var clip = ScriptableObject.CreateInstance<BlendShapeClip>();
+            clip.BlendShapeName = Path.GetFileNameWithoutExtension(path);
+            clip.Prefab = AssetDatabase.LoadAssetAtPath<GameObject>(AssetDatabase.GetAssetPath(this));
+            AssetDatabase.CreateAsset(clip, path);
+            AssetDatabase.ImportAsset(path);
+
+            Clips.Add(clip);
+            EditorUtility.SetDirty(this);
+            AssetDatabase.SaveAssets();
         }
 #endif
